python/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask==1.1.2
azure-storage-blob==12.17.0
Pillow==9.5.0